# compiler: namespace type equivalence based on AST node + captures