Sorry for the last commit here - I started work on this branch a while ago, and it's a big one so several conflicts accumulated, and I didn't quite resolve them right when rebasing.

---

This commit implements the accepted proposal #18816. Namespace-owning types (struct, enum, union, opaque) are now deduplicted based on a combination of their AST node and all values they capture. As well as comptime-known instructions, references to decls in outer namespaces which may be generic are now also treated as captures.

Reified types are deduplicated based on their AST node and the "structure" of the type. For instance, if two struct types are reified at the same source location and have identical layouts, fields, etc, they will be considered the same type. The way reified types work is a little odd at first glance - `Sema` constructs its own hash for the type which is used as the `InternPool` *key*. This is done primarily to simplify the `InternPool` key, since it would otherwise have to include all field types, aligns, default values, etc - this would make `Key` unnecessarily bulky, and would also make `indexToKey` more complex. This approach also means we can optimize for the case of the type already existing, by constructing this hash *without performing validation* (e.g. checking that field `alignment` is a power of two), and only performing the validation in a second pass constructing the type if it did not already exist.

This should unblock progress on incremental compilation in the frontend.