@andrewrk 2024-03-05
> Sorry for the last commit here - I started work on this branch a while ago, and it's a big one so several conflicts accumulated, and I didn't quite resolve them right when rebasing.

not a problem at all - that's why we do --no-ff merges when there is more than one commit :-)