@andrewrk 2024-03-05
Can we get a performance data point for this big boy?