@mlugg 2024-03-05
First, ZIR sizes - all insignificant improvements.

| File                | Old bytes   | New bytes   | Delta   |
|---------------------|-------------|-------------|---------|
| `Sema.zig`          | 4.04970 MiB | 4.04969 MiB | -0.00%  |
| `array_list.zig`    | 230.314 KiB | 230.295 KiB | -0.01%  |
| `mem.zig`           | 569.528 KiB | 569.426 KiB | -0.01%  |
| `meta.zig`          | 131.336 KiB | 131.297 KiB | -0.02%  |

Now for some performance benchmarks. Note that when I run the behavior tests
below, I'm running a subset of tests which pass on both `master` and this
branch, so the tests are identical across branches.

# AstGen - insignificant difference

```
Benchmark 1 (179 runs): ../master/stage4-fast/bin/zig ast-check src/Sema.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           112ms ± 5.95ms     108ms …  130ms         25 (14%)        0%
  peak_rss           62.5MB ± 35.8KB    62.3MB … 62.5MB         16 ( 9%)        0%
  cpu_cycles          324M  ± 1.04M      322M  …  326M           0 ( 0%)        0%
  instructions        522M  ± 85.2       522M  …  522M           1 ( 1%)        0%
  cache_references   9.18M  ±  205K     8.75M  … 10.3M           2 ( 1%)        0%
  cache_misses        641K  ± 75.0K      507K  …  783K           0 ( 0%)        0%
  branch_misses      3.76M  ± 9.22K     3.74M  … 3.78M           0 ( 0%)        0%
Benchmark 2 (180 runs): ./stage4-fast/bin/zig ast-check src/Sema.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           111ms ± 3.82ms     108ms …  129ms         12 ( 7%)          -  0.8% ±  0.9%
  peak_rss           62.4MB ± 65.5KB    62.2MB … 62.5MB          0 ( 0%)          -  0.1% ±  0.0%
  cpu_cycles          324M  ±  999K      321M  …  328M          30 (17%)          +  0.1% ±  0.1%
  instructions        521M  ± 78.9       521M  …  521M          45 (25%)          -  0.1% ±  0.0%
  cache_references   9.20M  ±  148K     8.74M  … 9.58M           6 ( 3%)          +  0.2% ±  0.4%
  cache_misses        645K  ± 44.9K      490K  …  709K          18 (10%)          +  0.6% ±  2.0%
  branch_misses      3.75M  ± 30.3K     3.71M  … 3.87M          37 (21%)          -  0.2% ±  0.1%
```

# Analyze behavior - performance improvement

If I had to guess, I'd say this comes from the elimination of `CaptureScope`. Aside from eliminating the overhead
of creating them etc, this change might have allowed `Sema.analyzeBodyInner` to tail call again!

```
Benchmark 1 (31 runs): ../master/stage4-fast/bin/zig test -Itest test/behavior.zig --test-no-exec -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           661ms ± 14.6ms     644ms …  727ms          1 ( 3%)        0%
  peak_rss            104MB ±  261KB     104MB …  105MB          0 ( 0%)        0%
  cpu_cycles         2.21G  ± 13.2M     2.19G  … 2.23G           0 ( 0%)        0%
  instructions       2.84G  ± 11.9K     2.84G  … 2.84G           0 ( 0%)        0%
  cache_references    276M  ± 1.90M      273M  …  280M           0 ( 0%)        0%
  cache_misses       37.9M  ± 1.26M     36.0M  … 40.3M           0 ( 0%)        0%
  branch_misses      14.4M  ± 91.9K     14.2M  … 14.5M           0 ( 0%)        0%
Benchmark 2 (32 runs): ./stage4-fast/bin/zig test -Itest test/behavior.zig --test-no-exec -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           643ms ± 6.44ms     631ms …  652ms          0 ( 0%)        ⚡-  2.8% ±  0.9%
  peak_rss            103MB ±  185KB     102MB …  103MB          0 ( 0%)        ⚡-  1.7% ±  0.1%
  cpu_cycles         2.13G  ± 13.1M     2.11G  … 2.15G           0 ( 0%)        ⚡-  3.7% ±  0.3%
  instructions       2.72G  ± 7.34K     2.72G  … 2.72G           1 ( 3%)        ⚡-  4.0% ±  0.0%
  cache_references    267M  ± 2.34M      263M  …  272M           0 ( 0%)        ⚡-  3.2% ±  0.4%
  cache_misses       40.6M  ± 1.56M     37.8M  … 44.6M           0 ( 0%)        💩+  7.1% ±  1.9%
  branch_misses      13.9M  ± 92.7K     13.8M  … 14.0M           0 ( 0%)        ⚡-  3.0% ±  0.3%
```

# Build behavior (x86 backend) - significant performance improvement

Honestly, I can't explain this one - I guess somehow one of my minor changes to the backend improved performance.

```
Benchmark 1 (11 runs): ../master/stage4-fast/bin/zig test -Itest test/behavior.zig --test-no-exec -fno-llvm -fno-lld
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.86s  ±  111ms    1.68s  … 2.09s           0 ( 0%)        0%
  peak_rss            120MB ±  330KB     119MB …  120MB          0 ( 0%)        0%
  cpu_cycles         4.98G  ± 27.0M     4.96G  … 5.04G           0 ( 0%)        0%
  instructions       6.54G  ± 24.9K     6.53G  … 6.54G           0 ( 0%)        0%
  cache_references    402M  ± 3.01M      399M  …  406M           0 ( 0%)        0%
  cache_misses       95.1M  ± 1.45M     93.5M  … 97.7M           0 ( 0%)        0%
  branch_misses      41.6M  ±  151K     41.5M  … 41.9M           0 ( 0%)        0%
Benchmark 2 (13 runs): ./stage4-fast/bin/zig test -Itest test/behavior.zig --test-no-exec -fno-llvm -fno-lld
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.66s  ± 60.1ms    1.58s  … 1.78s           0 ( 0%)        ⚡- 10.9% ±  4.0%
  peak_rss            117MB ±  175KB     117MB …  117MB          0 ( 0%)        ⚡-  2.4% ±  0.2%
  cpu_cycles         4.97G  ± 19.0M     4.93G  … 4.98G           4 (31%)          -  0.3% ±  0.4%
  instructions       6.47G  ± 21.4K     6.47G  … 6.47G           1 ( 8%)          -  1.0% ±  0.0%
  cache_references    416M  ± 2.29M      411M  …  419M           2 (15%)        💩+  3.5% ±  0.6%
  cache_misses       99.1M  ± 1.41M     95.9M  …  101M           2 (15%)        💩+  4.2% ±  1.3%
  branch_misses      41.9M  ±  113K     41.7M  … 42.0M           3 (23%)          +  0.8% ±  0.3%
```

# Analyze std tests - insignificant difference

```
Benchmark 1 (3 runs): ../master/stage4-fast/bin/zig test --zig-lib-dir lib lib/std/std.zig --test-no-exec -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          15.6s  ± 17.4ms    15.5s  … 15.6s           0 ( 0%)        0%
  peak_rss            405MB ±  205KB     404MB …  405MB          0 ( 0%)        0%
  cpu_cycles         54.8G  ± 45.1M     54.8G  … 54.9G           0 ( 0%)        0%
  instructions       83.2G  ± 5.22K     83.2G  … 83.2G           0 ( 0%)        0%
  cache_references   5.30G  ± 12.6M     5.29G  … 5.32G           0 ( 0%)        0%
  cache_misses        598M  ± 7.20M      593M  …  606M           0 ( 0%)        0%
  branch_misses       241M  ±  414K      240M  …  241M           0 ( 0%)        0%
Benchmark 2 (3 runs): ./stage4-fast/bin/zig test --zig-lib-dir lib lib/std/std.zig --test-no-exec -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          15.4s  ± 32.4ms    15.4s  … 15.5s           0 ( 0%)          -  0.8% ±  0.4%
  peak_rss            388MB ± 76.1KB     388MB …  388MB          0 ( 0%)        ⚡-  4.1% ±  0.1%
  cpu_cycles         54.5G  ±  131M     54.3G  … 54.6G           0 ( 0%)          -  0.6% ±  0.4%
  instructions       83.1G  ± 9.01K     83.1G  … 83.1G           0 ( 0%)          -  0.1% ±  0.0%
  cache_references   5.17G  ± 68.9M     5.12G  … 5.24G           0 ( 0%)          -  2.6% ±  2.1%
  cache_misses        654M  ± 31.3M      629M  …  689M           0 ( 0%)          +  9.3% ±  8.6%
  branch_misses       241M  ±  263K      241M  …  241M           0 ( 0%)          +  0.2% ±  0.3%
```